adds one test file collated
