--- sniffr/tests/funtional/homepage_test.py
@@ -1,12 +1,0 @@
-from sniffr.app import create_app
-
-def test_home_page_with_fixture(test_client):
-    """
-    GIVEN a Flask application configured for testing
-    WHEN the '/' page is requested (GET)
-    THEN check that the response is valid
-    """
-    response = test_client.get('/')
-    assert response.status_code == 200
-    assert b"Sniffr's" in response.data
-
