+++ tests/conftest.py
@@ -1,6 +1,5 @@
 from sniffr.app import create_app
 import pytest
-from sniffr.models import User
 from seed_db import seed_db
 
 @pytest.fixture(scope='module')
@@ -13,8 +12,3 @@ def test_client():
         with flask_app.app_context():
             seed_db()
             yield testing_client
-
-@pytest.fixture(scope='module')
-def new_user(test_client):
-    user = User('danny@d300.org', 'gancho')
-    return user
